[runbooks] Add Details Page and Opsgenie Integration
The runbooks plugin has a details page now, to view a runbook based on
the alert name and group.

The runbooks plugin can now also be used within the Opsgenie plugin, so
that runbooks can be added to Opsgenie alerts.
