+++ app/packages/runbooks/src/components/RunbooksPage.test.tsx
@@ -21,32 +21,46 @@ vi.mock('@kobsio/core', async () => {
 
 describe('RunbooksPage', () => {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const render = (): RenderResult => {
+  const render = (path: string): RenderResult => {
     const client = new APIClient();
     const getSpy = vi.spyOn(client, 'get');
-    getSpy.mockResolvedValue([
-      {
-        alert: 'test1',
-        expr: 'vector(1)',
-        group: 'testgroup1',
-        id: '/group/test/alert/test1',
-        message: 'Message 1',
-        runbook: 'Runbook 1',
-        severity: 'warning',
-      },
-      {
-        alert: 'test2',
-        expr: 'vector(2)',
-        group: 'testgroup2',
-        id: '/group/test/alert/test2',
-        message: 'Message 2',
-        runbook: 'Runbook 2',
-        severity: 'info',
-      },
-    ]);
+    if (path === '/') {
+      getSpy.mockResolvedValue([
+        {
+          alert: 'test1',
+          expr: 'vector(1)',
+          group: 'testgroup1',
+          id: '/group/test/alert/test1',
+          message: 'Message 1',
+          runbook: 'Runbook 1',
+          severity: 'warning',
+        },
+        {
+          alert: 'test2',
+          expr: 'vector(2)',
+          group: 'testgroup2',
+          id: '/group/test/alert/test2',
+          message: 'Message 2',
+          runbook: 'Runbook 2',
+          severity: 'info',
+        },
+      ]);
+    } else {
+      getSpy.mockResolvedValue([
+        {
+          alert: 'test1',
+          expr: 'vector(1)',
+          group: 'testgroup1',
+          id: '/group/test/alert/test1',
+          message: 'Message 1',
+          runbook: 'Runbook 1',
+          severity: 'warning',
+        },
+      ]);
+    }
 
     return _render(
-      <MemoryRouter>
+      <MemoryRouter initialEntries={[path]}>
         <QueryClientProvider>
           <APIContext.Provider value={{ client: client, getUser: () => undefined }}>
             <RunbooksPage
@@ -64,7 +78,7 @@ describe('RunbooksPage', () => {
   };
 
   it('should render runbooks', async () => {
-    render();
+    render('/');
 
     expect(screen.getByText('runbooks')).toBeInTheDocument();
     expect(screen.getByText('(hub / runbooks)')).toBeInTheDocument();
@@ -81,6 +95,20 @@ describe('RunbooksPage', () => {
     expect(await waitFor(() => screen.getByText('Message 2'))).toBeInTheDocument();
 
     await userEvent.click(screen.getByText(/test1/));
+    expect(await waitFor(() => screen.getByText('Runbook 1'))).toBeInTheDocument();
+  });
+
+  it('should render runbook', async () => {
+    render('/group/testgroup1/alert/test1');
+
+    expect(screen.getByText('runbooks')).toBeInTheDocument();
+    expect(screen.getByText('(hub / runbooks)')).toBeInTheDocument();
+    expect(screen.getByText(description)).toBeInTheDocument();
+
+    expect(await waitFor(() => screen.getByText('test1'))).toBeInTheDocument();
+    expect(await waitFor(() => screen.getByText('testgroup1'))).toBeInTheDocument();
+    expect(await waitFor(() => screen.getByText('warning'))).toBeInTheDocument();
+    expect(await waitFor(() => screen.getByText('Message 1'))).toBeInTheDocument();
     expect(await waitFor(() => screen.getByText('Runbook 1'))).toBeInTheDocument();
   });
 });
