+++ app/packages/runbooks/src/components/RunbooksPanel.test.tsx
@@ -22,26 +22,43 @@ describe('RunbooksPanel', () => {
   const render = (options: any): RenderResult => {
     const client = new APIClient();
     const getSpy = vi.spyOn(client, 'get');
-    getSpy.mockResolvedValue([
-      {
-        alert: 'test1',
-        expr: 'vector(1)',
-        group: 'testgroup1',
-        id: '/group/test/alert/test1',
-        message: 'Message 1',
-        runbook: 'Runbook 1',
-        severity: 'warning',
-      },
-      {
-        alert: 'test2',
-        expr: 'vector(2)',
-        group: 'testgroup2',
-        id: '/group/test/alert/test2',
-        message: 'Message 2',
-        runbook: 'Runbook 2',
-        severity: 'info',
-      },
-    ]);
+
+    if (options?.type === 'list') {
+      getSpy.mockResolvedValue([
+        {
+          alert: 'test1',
+          expr: 'vector(1)',
+          group: 'testgroup1',
+          id: '/group/test/alert/test1',
+          message: 'Message 1',
+          runbook: 'Runbook 1',
+          severity: 'warning',
+        },
+        {
+          alert: 'test2',
+          expr: 'vector(2)',
+          group: 'testgroup2',
+          id: '/group/test/alert/test2',
+          message: 'Message 2',
+          runbook: 'Runbook 2',
+          severity: 'info',
+        },
+      ]);
+    } else if (options?.type === 'details') {
+      getSpy.mockResolvedValue([
+        {
+          alert: 'test1',
+          expr: 'vector(1)',
+          group: 'testgroup1',
+          id: '/group/test/alert/test1',
+          message: 'Message 1',
+          runbook: 'Runbook 1',
+          severity: 'warning',
+        },
+      ]);
+    } else {
+      getSpy.mockResolvedValue([]);
+    }
 
     return _render(
       <MemoryRouter>
@@ -76,8 +93,8 @@ describe('RunbooksPanel', () => {
     expect(await waitFor(() => screen.getByText('Invalid options for Runbooks plugin'))).toBeInTheDocument();
   });
 
-  it('should render list of runbooks releases', async () => {
-    render({ query: 'test' });
+  it('should render list of runbooks', async () => {
+    render({ query: 'test', type: 'list' });
 
     expect(await waitFor(() => screen.getByText('test1'))).toBeInTheDocument();
     expect(await waitFor(() => screen.getByText('testgroup1'))).toBeInTheDocument();
@@ -90,6 +107,16 @@ describe('RunbooksPanel', () => {
     expect(await waitFor(() => screen.getByText('Message 2'))).toBeInTheDocument();
 
     await userEvent.click(screen.getByText(/test1/));
+    expect(await waitFor(() => screen.getByText('Runbook 1'))).toBeInTheDocument();
+  });
+
+  it('should render single runbook', async () => {
+    render({ alert: 'test1', group: 'testgroup1', type: 'details' });
+
+    expect(await waitFor(() => screen.getByText('test1'))).toBeInTheDocument();
+    expect(await waitFor(() => screen.getByText('testgroup1'))).toBeInTheDocument();
+    expect(await waitFor(() => screen.getByText('warning'))).toBeInTheDocument();
+    expect(await waitFor(() => screen.getByText('Message 1'))).toBeInTheDocument();
     expect(await waitFor(() => screen.getByText('Runbook 1'))).toBeInTheDocument();
   });
 });
